+++ src/Api/Projects.php
@@ -233,6 +233,23 @@ class Projects extends AbstractApi
 
     /**
      * @param int|string $project_id
+     * @param string     $ref
+     * @param string     $token
+     * @param array      $variables
+     *
+     * @return mixed
+     */
+    public function triggerPipeline($project_id, string $ref, string $token, array $variables = [])
+    {
+        return $this->post($this->getProjectPath($project_id, 'trigger/pipeline'), [
+            'ref' => $ref,
+            'token' => $token,
+            'variables' => $variables,
+        ]);
+    }
+
+    /**
+     * @param int|string $project_id
      * @param array      $parameters {
      *
      *     @var string $scope       the scope of pipelines, one of: running, pending, finished, branches, tags